build: Fix the release workflow and tests workflow triggers (#96)
* build: first version of automatic release script

* build: add option for running test PyPI and edit input names

* build: uncapitalize the release workflow name

* ci: only run tests workflow on push for the main branch

Currently, all ci tests are run twice in a pull request, both for the push and the pull request event. This fix prevents this.

* build: update input argument descriptions

* build: Don't run make docs

Not necessary in the automated build process. Used in the semi-automated one earlier for visual verification.

* build: remove skchange install

Not needed anymore, since the docs are not built.

* build: install twine in every job

* build: publish the github release as a draft

For now. Until tested for a couple of versions.

* build: add license info

Attempt to fix InvalidDistribution on TestPyPI

* build: restrict twine to <= 6.0.1

https://stackoverflow.com/questions/79408101/what-is-the-correct-way-of-specifying-the-license-in-pyproject-toml-file-for-a-n

* build: run twine upload with verbose option for TestPyPI

* build: revert add license info commit

* build: revert to input args dry-run and skip-test-pypi in release script